changed price to float

+++ db/migrate/20240511212107_add_price_to_store_items.rb
@@ -1,5 +1,5 @@
 class AddPriceToStoreItems < ActiveRecord::Migration[7.1]
   def change
-    add_column :store_items, :price, :integer
+    add_column :store_items, :price, :float
   end
 end
